[Update] detach calendar initialization module and modify likes data update module
