+++ README.md
@@ -102,7 +102,8 @@
 
 ### _2022.01.16 (토)_
 - 좋아요 클릭시 useMutation 으로 서버에 있는 feed 데이터 업데이트
-- scroll css 수정
+    - 근데 취소했을때 서버에 요청은 오는데 반영이 안됨...??? 일단 푸시함.
+- scroll css 수정 및 달력 초기화 모듈 분리
 
 ## 태양
 ---
