+++ src/components/views/community/Feed.tsx
@@ -12,11 +12,10 @@ import { useMutation } from "react-query";
 import axios from "axios";
 
 const Feed:React.FC<Post> = ({feed}:Post) => {
-
+    const [init, setInit] = useState<boolean>(false);
     const [showCmts, setShowCmts] = useState<boolean>(false);
-    const [initData, setInitData] = useState<boolean>(false);
     const [reply, setReply] = useState<Comment>();
-    const [likes, setLikes] = useState<number>(0);
+    const [likes, setLikes] = useState<number | undefined>(undefined);
     
     const mutation = useMutation( async () => {
         axios.patch(`api/community/post/edit/${feed.id}/${likes}`);
@@ -33,11 +32,11 @@ const Feed:React.FC<Post> = ({feed}:Post) => {
         };
         feed.comments.push(data);
         setReply(data);
-        setInitData(true);
+        setInit(true);
     },[feed]);
 
     useEffect(() => {
-        if(initData && feed.likes !== likes) {
+        if(likes !== undefined && feed.likes !== likes && init) {
             mutation.mutateAsync().then(() => console.log('update number of likes'));
         }
     },[likes]);
@@ -56,7 +55,7 @@ const Feed:React.FC<Post> = ({feed}:Post) => {
                     <Icons>
                         <Comments cnt={feed.comments.length - 1} showCmts={showCmts} setShowCmts={setShowCmts}/>
                         {/* <Like /> */}
-                        <Heart likes={likes} setLikes={setLikes}/>
+                        <Heart likes={likes === undefined ? 0 : likes} setLikes={setLikes}/>
                         <ShareButton />
                     </Icons>
                 </CardPadding>
