+++ src/components/views/community/calendar.tsx
@@ -2,7 +2,7 @@
 import { css, keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
 import React, { useEffect, useState } from 'react';
-import { weekdays, months } from './CalendarBase';
+import { weekdays, months, init } from './CalendarBase';
 import { DateProps as Props } from '../../../types/components/views/DateProps';
 import { NextBtn, PrevBtn } from './CalendarBtn';
 import { leftPad } from '../../../lib/common/DateUtils';
@@ -17,41 +17,13 @@ const Calendar: React.FC<Props> = (props: Props) => {
   useEffect(() => {
     setYear(props.year);
     setMonth(props.month);
-    init(year, month);
+    setDateSet(init(year, month));
   }, []);
 
   useEffect(() => {
-    init(year, month);
+    setDateSet(init(year, month));
   }, [year, month]);
 
-  const init = (year: number, month: number) => {
-    const start: Date = new Date(year, month, 1);
-    const end: Date = new Date(year, month + 1, 1 - 1);
-
-    let startDayOfWeek: number = start.getDay();
-    let endDayOfWeek: number = end.getDay();
-    let endDate: number = end.getDate();
-
-    let rows: number[][] = [];
-    let days: number[] = [];
-
-    for (let i = 0; i < startDayOfWeek; i++) {
-      days.push(0);
-    }
-    for (let i = 1, j = startDayOfWeek; i <= endDate; i++, j++) {
-      if (j % 7 === 6) {
-        days.push(i);
-        rows.push(days);
-        days = [];
-      } else days.push(i);
-    }
-    for (let i = endDayOfWeek; i < 6; i++) {
-      days.push(0);
-    }
-    rows.push(days);
-
-    setDateSet(rows);
-  };
 
   return (
     <>
